Added test data to the test for archive update and logger reconstruction.

+++ code-preprocessing/archive-update/test_archives.py
@@ -1,7 +1,9 @@
 # A series of tests to check whether the python scripts of archive-update perform correctly.
 # Start the tests by writing
 # py.test
-# in a terminal window
+# or
+# python -m pytest
+# in a terminal window on this folder
 
 from os.path import dirname, abspath, join, exists
 from os import walk, remove, rmdir, chdir, chmod
@@ -11,27 +13,28 @@ def almost_equal(value1, value2, precision):
     return abs(value1 - value2) < precision
 
 
-def prepare_archive_data():
+def prepare_archive_data(download_data=False):
     """
-    Unpacks the archive data to the test-data folder.
+    Prepares the data needed for the tests (cleans up the test-data folder of unnecessary files) and, if download_data
+    is True, downloads the test data from the internet.
     """
     import urllib
     import tarfile
     cleanup_archive_data()
     data_folder = abspath(join(dirname(__file__), 'test-data'))
-    if not exists(abspath(join(data_folder, 'archives-input'))) or not exists(
-            abspath(join(data_folder, 'archives-results'))):
+    if download_data and (not exists(abspath(join(data_folder, 'archives-input'))) or not exists(
+            abspath(join(data_folder, 'archives-results')))):
         cleanup_archive_data(True)
         chdir(abspath(dirname(__file__)))
-        data_url = 'http://dis.ijs.si/tea/tmp/archive-update-test-data.tgz'
+        data_url = 'link-to-archive-update-test-data.tgz'
         filename, headers = urllib.urlretrieve(data_url)
         tar_file = tarfile.open(filename)
         tar_file.extractall()
 
-    for root, dirs, files in walk(data_folder, topdown=False):
-        for name in files:
-            # Change file permission so it can be deleted
-            chmod(join(root, name), 0777)
+        for root, dirs, files in walk(data_folder, topdown=False):
+            for name in files:
+                # Change file permission so it can be deleted
+                chmod(join(root, name), 0777)
 
 
 def cleanup_archive_data(delete_all=False):
